requirements.txt
python-docx>=0.8.11
# 其他可能的依赖
# pandas>=1.5.0
# numpy>=1.24.0